[tests] Fix function signature (mono/mono#13090)
[tests] Fix function signature

To match pinvoke signature from pinvoke-2 test.


Commit migrated from https://github.com/mono/mono/commit/809e40435f9559cd2c8c130a07f491880ee1b0ce

+++ src/mono/mono/tests/libtest.c
@@ -130,7 +130,7 @@ mono_cominterop_is_supported (void)
 }
 
 LIBTEST_API unsigned short* STDCALL
-test_lpwstr_marshal (unsigned short* chars, long length)
+test_lpwstr_marshal (unsigned short* chars, int length)
 {
 	int i = 0;
 	unsigned short *res;
